// frontend/lib/generated/fastapi/schemas/browseProductsPageResult.ts
/**
 * Generated by orval v7.20.0 🍺
 * Do not edit manually.
 * ShipyardHQ API
 * OpenAPI spec version: 0.1.0
 */
import type { PublicProductCard } from './publicProductCard';

export interface BrowseProductsPageResult {
  items: PublicProductCard[];
  hasMore: boolean;
  page: number;
  pageSize: number;
  total: number;
}